use more aggressive hex encoding instead of url encoding for path format
Rolling with this for now due to some insane URL decoding in older
versions of node (<= v0.2.4). Ostensibly, old node versions decode
the piss out of anything in the path part of the URL, include %2F
and also squash // characters together. The result is that you cannot
cleanly separate a URL encoded value in the path.

With this new scheme, every byte of the image URL is encoded into a
2 char hex value. I actually kind of like it for its similarity to
the digest hex encoding.