Don't write a PID file in `pwd`/tmp/
If `camo` is running as a privileged user the current behaviour would
allow a race condition leading to the overwriting of arbitrary files.

This is exploitable when running `camo` from an init script, since then
your process as a working directory of `/`.

Since `camo` does not fork into the background there isn't really a reason
to write its PID out to a file; people writing initscripts can simply use
`start-stop-daemon`'s `--make-pidfile` flag.