Fix socket timeout
Change the log message and variable names because Socket.setTimeout is not
a connection timeout, but a timeout after inactivity on the socket.

It is also necesary to close the connection with `.abort()` rather than
`.end()`. end only signals the end of the request body and does not close
the connection.